requirements.txt
Django==3.2.7
django-import-export==2.7.1
python-decouple==3.5
pythonping==1.1.0